producer/requirements.txt
kafka-python==2.0.2
prometheus-client==0.14.1
structlog==21.5.0
python-json-logger==2.0.4
jsonschema==4.17.3
retry==0.9.2
flask==2.2.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0
python-snappy==0.6.1
